% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-plot.R
\name{ispat3d_plot_zones}
\alias{ispat3d_plot_zones}
\title{Facet all fitted zone networks in one base-R figure}
\usage{
ispat3d_plot_zones(fit, zones = names(fit$full), columns = 2L, ...)
}
\arguments{
\item{fit}{An ISPAT3D fit from ispat3d_fit() or ispat3d_fit_2d().}

\item{zones}{Zone names to plot, in order.}

\item{columns}{Number of panel columns.}

\item{...}{Further arguments passed to ispat3d_plot_network().}
}
\value{
Invisibly returns a named list of displayed edge tables.
}
\description{
Facet all fitted zone networks in one base-R figure
}
